Correctly handle string table names with fallback langs
Prior to this change, if you passed a custom table name and used a
language where a fallback language was used to determine the
pluralization rule, the tableName variable would be passed to [self
pathForResource:...] multiple times. On the second time, passing an
absolute string would produce a nil result.

This would happen, for example, for Mexican Spanish, where the
lang list is ("es_MX", "es-MX", "es", "English"), and "es"
determines the pluralization rule.